Added reStructCategories as a helper function
